requirements.txt
python-jose[cryptography]
aio-pika==9.5.5
aiopg==1.4.0
aiormq==6.8.1
alembic==1.16.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
asyncpg==0.30.0
billiard==4.2.1
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.3
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.12
greenlet==3.2.3
h11==0.16.0
idna==3.10
kombu==5.5.4
Mako==1.3.10
MarkupSafe==3.0.2
multidict==6.4.4
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
redis==6.2.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uvicorn==0.34.3
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0